refactor exception packages
